add readme and fix code
